rename controller files
