+++ src/libs/extension-api/controller/base-extensions-initializer.test.ts
@@ -2,7 +2,7 @@ import { expect, fixture } from '@open-wc/testing';
 import { UmbExtensionRegistry } from '../registry/extension.registry.js';
 import { ManifestCondition, ManifestWithDynamicConditions, UmbConditionConfigBase } from '../types.js';
 import { UmbExtensionCondition } from '../condition/extension-condition.interface.js';
-import { PermittedControllerType, UmbBaseExtensionController, UmbBaseExtensionsController } from './index.js';
+import { PermittedControllerType, UmbBaseExtensionInitializer, UmbBaseExtensionsInitializer } from './index.js';
 import {
 	UmbBaseController,
 	UmbControllerHost,
@@ -13,7 +13,7 @@ import { customElement, html } from '@umbraco-cms/backoffice/external/lit';
 @customElement('umb-test-controller-host')
 class UmbTestControllerHostElement extends UmbControllerHostElementMixin(HTMLElement) {}
 
-class UmbTestExtensionController extends UmbBaseExtensionController {
+class UmbTestExtensionController extends UmbBaseExtensionInitializer {
 	constructor(
 		host: UmbControllerHost,
 		extensionRegistry: UmbExtensionRegistry<ManifestWithDynamicConditions>,
@@ -41,7 +41,7 @@ type myTestManifestTypes = myTestManifestTypesUnion | myTestManifestTypesUnion[]
 
 class UmbTestExtensionsController<
 	MyPermittedControllerType extends UmbTestExtensionController = PermittedControllerType<UmbTestExtensionController>
-> extends UmbBaseExtensionsController<
+> extends UmbBaseExtensionsInitializer<
 	myTestManifests,
 	myTestManifestTypesUnion,
 	ManifestWithDynamicConditions,
