+++ src/libs/extension-api/controller/extension-api-initializer.test.ts
@@ -1,31 +1,43 @@
 import { expect, fixture } from '@open-wc/testing';
 import { UmbExtensionRegistry } from '../registry/extension.registry.js';
-import { UmbExtensionElementController } from './index.js';
-import { UmbControllerHostElement, UmbControllerHostElementMixin } from '@umbraco-cms/backoffice/controller-api';
+import { ManifestApi, ManifestWithDynamicConditions } from '../types.js';
+import { UmbExtensionApiInitializer } from './index.js';
+import { UmbBaseController, UmbControllerHost, UmbControllerHostElement, UmbControllerHostElementMixin } from '@umbraco-cms/backoffice/controller-api';
 import { customElement, html } from '@umbraco-cms/backoffice/external/lit';
 import { type ManifestSection, UmbSwitchCondition } from '@umbraco-cms/backoffice/extension-registry';
 
 @customElement('umb-test-controller-host')
 export class UmbTestControllerHostElement extends UmbControllerHostElementMixin(HTMLElement) {}
 
-describe('UmbExtensionElementController', () => {
+
+export class UmbTestApiController extends UmbBaseController {
+
+	public i_am_test_api_controller = true;
+
+	constructor(host:UmbControllerHost) {
+		super(host);
+	}
+
+}
+
+interface TestManifest extends ManifestWithDynamicConditions, ManifestApi<UmbTestApiController> {
+	type: 'test-type'
+}
+
+describe('UmbExtensionApiController', () => {
 	describe('Manifest without conditions', () => {
 		let hostElement: UmbControllerHostElement;
-		let extensionRegistry: UmbExtensionRegistry<ManifestSection>;
-		let manifest: ManifestSection;
+		let extensionRegistry: UmbExtensionRegistry<TestManifest>;
+		let manifest: TestManifest;
 
 		beforeEach(async () => {
 			hostElement = await fixture(html`<umb-test-controller-host></umb-test-controller-host>`);
 			extensionRegistry = new UmbExtensionRegistry();
 			manifest = {
-				type: 'section',
-				name: 'test-section-1',
-				alias: 'Umb.Test.Section.1',
-				elementName: 'section',
-				meta: {
-					label: 'my section',
-					pathname: 'my-section',
-				},
+				type: 'test-type',
+				name: 'test-type-1',
+				alias: 'Umb.Test.Type-1',
+				api: UmbTestApiController
 			};
 
 			extensionRegistry.register(manifest);
@@ -33,50 +45,45 @@ describe('UmbExtensionElementController', () => {
 
 		it('permits when there is no conditions', (done) => {
 			let called = false;
-			const extensionController = new UmbExtensionElementController(
+			const extensionController = new UmbExtensionApiInitializer<TestManifest>(
 				hostElement,
 				extensionRegistry,
-				'Umb.Test.Section.1',
+				'Umb.Test.Type-1',
+				[hostElement],
 				(permitted) => {
 					if (called === false) {
 						called = true;
 						expect(permitted).to.be.true;
 						if (permitted) {
-							expect(extensionController?.manifest?.alias).to.eq('Umb.Test.Section.1');
-							expect(extensionController.component?.nodeName).to.eq('SECTION');
+							expect(extensionController?.manifest?.alias).to.eq('Umb.Test.Type-1');
+							expect(extensionController.api?.i_am_test_api_controller).to.be.true;
 							done();
 							extensionController.destroy();
 						}
 					}
 				}
 			);
-		});
-
-		it('utilized the default element when there is none provided by manifest', (done) => {
-			extensionRegistry.unregister(manifest.alias);
-
-			const noElementManifest = { ...manifest, elementName: undefined };
-			extensionRegistry.register(noElementManifest);
-
-			let called = false;
-			const extensionController = new UmbExtensionElementController(
+			/*
+			TODO: Consider if builder pattern would be a more nice way to setup this:
+			const extensionController = new UmbExtensionApiController<TestManifest>(
 				hostElement,
 				extensionRegistry,
-				'Umb.Test.Section.1',
-				(permitted) => {
-					if (called === false) {
-						called = true;
-						expect(permitted).to.be.true;
-						if (permitted) {
-							expect(extensionController?.manifest?.alias).to.eq('Umb.Test.Section.1');
-							expect(extensionController.component?.nodeName).to.eq('UMB-TEST-FALLBACK-ELEMENT');
-							done();
-							extensionController.destroy();
-						}
+				'Umb.Test.Type-1'
+			)
+			.withConstructorArguments([hostElement])
+			.onPermitted((permitted) => {
+				if (called === false) {
+					called = true;
+					expect(permitted).to.be.true;
+					if (permitted) {
+						expect(extensionController?.manifest?.alias).to.eq('Umb.Test.Type-1');
+						expect(extensionController.api?.i_am_test_api_controller).to.be.true;
+						done();
+						extensionController.destroy();
 					}
-				},
-				'umb-test-fallback-element'
-			);
+				}
+			).observe();
+			*/
 		});
 	});
 
@@ -90,10 +97,10 @@ describe('UmbExtensionElementController', () => {
 			extensionRegistry = new UmbExtensionRegistry();
 
 			manifest = {
-				type: 'section',
-				name: 'test-section-1',
-				alias: 'Umb.Test.Section.1',
-				elementName: 'section',
+				type: 'test-type',
+				name: 'test-type-1',
+				alias: 'Umb.Test.Type-1',
+				api: UmbTestApiController,
 				conditions: [
 					{
 						alias: 'Umb.Test.Condition.Delay',
@@ -125,21 +132,22 @@ describe('UmbExtensionElementController', () => {
 
 		it('does change permission as conditions change', (done) => {
 			let count = 0;
-			const extensionController = new UmbExtensionElementController(
+			const extensionController = new UmbExtensionApiInitializer<TestManifest>(
 				hostElement,
 				extensionRegistry,
-				'Umb.Test.Section.1',
+				'Umb.Test.Type-1',
+				[hostElement],
 				async () => {
 					count++;
 					// We want the controller callback to first fire when conditions are initialized.
 					expect(extensionController.manifest?.conditions?.length).to.be.equal(2);
-					expect(extensionController?.manifest?.alias).to.eq('Umb.Test.Section.1');
+					expect(extensionController?.manifest?.alias).to.eq('Umb.Test.Type-1');
 					if (count === 1) {
 						expect(extensionController?.permitted).to.be.true;
-						expect(extensionController.component?.nodeName).to.eq('SECTION');
+						expect(extensionController.api?.i_am_test_api_controller).to.be.true;
 					} else if (count === 2) {
 						expect(extensionController?.permitted).to.be.false;
-						expect(extensionController.component).to.be.undefined;
+						expect(extensionController.api).to.be.undefined;
 						done();
 						extensionController.destroy(); // need to destroy the controller.
 					}
