+++ src/libs/extension-api/controller/index.ts
@@ -1,8 +1,8 @@
-export * from './base-extension-controller.js';
-export * from './base-extensions-controller.js';
-export * from './extension-api-controller.js';
-export * from './extensions-api-controller.js';
-export * from './extension-element-controller.js';
-export * from './extensions-element-controller.js';
-export * from './extension-manifest-controller.js';
-export * from './extensions-manifest-controller.js';
+export * from './base-extension-initializer.controller.js';
+export * from './base-extensions-initializer.controller.js';
+export * from './extension-api-initializer.controller.js';
+export * from './extensions-api-initializer.controller.js';
+export * from './extension-element-initializer.controller.js';
+export * from './extensions-element-initializer.controller.js';
+export * from './extension-manifest-initializer.controller.js';
+export * from './extensions-manifest-initializer.controller.js';
