+++ src/packages/core/action/repository-action.ts
@@ -1,5 +1,5 @@
 import { UmbBaseController, UmbControllerHostElement } from '@umbraco-cms/backoffice/controller-api';
-import { type UmbApi, UmbExtensionApiController } from '@umbraco-cms/backoffice/extension-api';
+import { type UmbApi, UmbExtensionApiInitializer } from '@umbraco-cms/backoffice/extension-api';
 import { umbExtensionsRegistry } from '@umbraco-cms/backoffice/extension-registry';
 
 export class UmbActionBase<RepositoryType> extends UmbBaseController implements UmbApi {
@@ -8,7 +8,7 @@ export class UmbActionBase<RepositoryType> extends UmbBaseController implements 
 	constructor(host: UmbControllerHostElement, repositoryAlias: string) {
 		super(host);
 
-		new UmbExtensionApiController(this, umbExtensionsRegistry, repositoryAlias, [this._host], (permitted, ctrl) => {
+		new UmbExtensionApiInitializer(this, umbExtensionsRegistry, repositoryAlias, [this._host], (permitted, ctrl) => {
 			this.repository = permitted ? ctrl.api as RepositoryType : undefined;
 		});
 		/*this.observe(
