+++ src/packages/core/components/extension-slot/extension-slot.element.ts
@@ -1,7 +1,7 @@
 import { type ManifestTypes, umbExtensionsRegistry } from '../../extension-registry/index.js';
 import { css, repeat, customElement, property, state, TemplateResult } from '@umbraco-cms/backoffice/external/lit';
 import {
-	type UmbExtensionElementController,
+	type UmbExtensionElementInitializer,
 	UmbExtensionsElementController,
 } from '@umbraco-cms/backoffice/extension-api';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
@@ -23,7 +23,7 @@ export class UmbExtensionSlotElement extends UmbLitElement {
 	#extensionsController?: UmbExtensionsElementController<ManifestTypes>;
 
 	@state()
-	private _permittedExts: Array<UmbExtensionElementController> = [];
+	private _permittedExts: Array<UmbExtensionElementInitializer> = [];
 
 	/**
 	 * The type or types of extensions to render.
@@ -95,7 +95,7 @@ export class UmbExtensionSlotElement extends UmbLitElement {
 	public defaultElement?:string;
 
 	@property()
-	public renderMethod?: (extension: UmbExtensionElementController) => TemplateResult | HTMLElement | null | undefined;
+	public renderMethod?: (extension: UmbExtensionElementInitializer) => TemplateResult | HTMLElement | null | undefined;
 
 	connectedCallback(): void {
 		super.connectedCallback();
